requirements.txt
tqdm          == 4.65.0
numpy         == 1.24.2
opencv-python == 4.7.0.72
pandas        == 2.0.0
torch
torchvision
transformers
scipy
python-highcharts
imageio
gdown
sk-video
pillow
moviepy
functools
pyiqa @ git+https://github.com/chaofengc/IQA-PyTorch.git
